requirements.txt
datasets>=2.20.0
faiss-cpu>=1.8.0
langchain>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=3.0.1
streamlit>=1.38.0
numpy>=1.26.4
scikit-learn>=1.4.2
tqdm>=4.66.4
python-dotenv>=1.0.1
google-generativeai>=0.7.2
openai>=1.44.0
tenacity>=8.4.2


